Created Summary object
Now instead of returning the JSON Summarizer returns a Summary object.

The Summary object is a Traversable[String], meaning you can iterate on
the sentences directly from it.
It also has two utilitiess: charCount lazy val and takeChars method. The
latter can be used to return a new Summary which obeys to the
limitCharCount passed as paremeter.